+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/realm/repository/RealmMessageRepository.kt
@@ -1,10 +1,13 @@
-package io.getstream.chat.android.offline.repository.domain.message.internal
+package io.getstream.chat.android.offline.repository.realm.repository
 
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.persistance.repository.MessageRepository
 import io.getstream.chat.android.client.query.pagination.AnyChannelPaginationRequest
 import io.getstream.chat.android.client.utils.SyncStatus
+import io.getstream.chat.android.offline.repository.realm.entity.MessageEntityRealm
+import io.getstream.chat.android.offline.repository.realm.entity.toDomain
+import io.getstream.chat.android.offline.repository.realm.entity.toRealm
 import io.realm.kotlin.Realm
 import io.realm.kotlin.UpdatePolicy
 import io.realm.kotlin.ext.query
